+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Change history of the project. All the feature updates, bug fixes, breaking chan
 ## [ 0.6.0 ]
 
 - Plugin signed
+- JSONPath support for JSON root selector
+- Support for Unix/epoch seconds format
 
 ## [ 0.5.0 ]